Moving issuance views to issuer
